+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable, unreleased changes to this project will be documented in this file.
 
 - Fix problem with free shipping voucher - #4942 by @IKarbowiak
 - Add sub-categories to random data - #4949 by @IKarbowiak
+- Deprecate `localized` field in Money type - #4952 by @IKarbowiak
+- Fix for shipping api doesn't apply taxes - #4913 by @kswiatek92
 - Allow to delete category and leave products - #4970 by @IKarbowiak
 
 ## 2.9.0